core/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
fake-useragent==1.2.1
frozenlist==1.4.0
idna==3.4
loguru==0.7.2
multidict==6.0.4
psycopg2==2.9.7
psycopg2-binary==2.9.7
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.5
yarl==1.9.2
